Change default quotes from double to single
- Updated default value in src/lib/core/questions.ts
- Updated fallback value in templates/prettier.config.mjs.ts
- Updated test baseAnswers to use single quotes
- Explicitly set double quotes for test that requires it
- Regenerated baseline test files with new default

+++ src/lib/core/questions.ts
@@ -780,7 +780,7 @@ export const questionGroups: QuestionGroup[] = [
 				name: "quotes",
 				label: "Quotes",
 				message: "Do you prefer double or single quotes?",
-				initial: "double",
+				initial: "single",
 				choices: ["double", "single"],
 				migrate: async ctx => ((await ctx.analyzeCode('"', "'")) ? "double" : "single"),
 			},
